+++ components-ui/app/src/test/filtering/FilteringUtils.test.tsx
@@ -20,7 +20,7 @@ import {
   removeComponent,
   addSubFilter,
   resetZone,
-  setFilter,
+  setFilterInput,
   filterListener,
   IZone,
   IFilter,
@@ -339,51 +339,6 @@ describe("Testing generateFilter function", () => {
     const testCall1 = generateFilter(mockZone, "component1", true);
     expect(testCall1).toEqual(combinedFilterValue);
   });
-
-  test("Generates filter correctly with ignored attributes", () => {
-    const filterValue1: IFilter = {
-      and_: {
-        attribute1: {
-          op1: {
-            value: 10,
-            negate: false,
-          },
-        },
-        attribute2: {
-          op2: {
-            value: 20,
-            negate: false,
-          },
-        },
-      },
-    };
-
-    const mockZone: IZone = {
-      components: {
-        component1: {
-          data: {
-            filter: filterValue1,
-          },
-        },
-      },
-      order: ["component1"],
-      type: "dashboard",
-    };
-
-    const expectedFilter: IFilter = {
-      and_: {
-        attribute2: {
-          op2: {
-            value: 20,
-            negate: false,
-          },
-        },
-      },
-    };
-
-    const testCall1 = generateFilter(mockZone, "component1", false, ["attribute1"]);
-    expect(testCall1).toEqual(expectedFilter);
-  });
 });
 
 describe("Testing addValueBelow function", () => {
@@ -635,7 +590,7 @@ describe("Testing setFilter function", () => {
       },
     };
 
-    setFilter({
+    setFilterInput({
       componentId: "component1",
       value: 10,
       zone: mockZone,
@@ -661,7 +616,7 @@ describe("Testing setFilter function", () => {
       type: "dashboard",
     };
 
-    setFilter({
+    setFilterInput({
       componentId: "component1",
       value: 10,
       zone: mockZone,
@@ -699,7 +654,7 @@ describe("Testing setFilter function", () => {
       },
     };
 
-    setFilter({
+    setFilterInput({
       componentId: "component1",
       value: 10,
       zone: mockZone,
@@ -736,7 +691,7 @@ describe("Testing setFilter function", () => {
       },
     };
 
-    setFilter({
+    setFilterInput({
       componentId: "component1",
       value: 10,
       zone: mockZone,
@@ -767,7 +722,7 @@ describe("Testing setFilter function", () => {
       type: "dashboard",
     };
 
-    setFilter({
+    setFilterInput({
       componentId: "component1",
       value: 10,
       zone: mockZone,