+++ components-ui/app/src/tol-ui/src/filtering/FilterMultiSelect.tsx
@@ -7,7 +7,7 @@ SPDX-License-Identifier: MIT
 import { useEffect, useState } from "react";
 import {
   IFilterInput,
-  setFilter,
+  setFilterInput,
   filterListener,
   MultipleSelect,
   FilterToggle,
@@ -17,8 +17,9 @@ import {
   TFilterOrUndefined,
   generateFilter,
   filterHasUpdated,
-  removeAttributeFromFilter,
   API_OPERATIONS,
+  resetFiltersBelow,
+  useEffectUpdate,
 } from "..";
 
 
@@ -34,32 +35,40 @@ export function FilterMultiSelect(props: IFilterInput) {
   } = props;
   const [data, setData] = useState<string[]>([]);
   const [values, setValues] = useState<string[]>([]);
-  const [disabled, setDisabled] = useState(false);
   const [exists, setExists] = useState<boolean>(false);
   const [negate, setNegate] = useState<boolean>(false);
   const [timeoutValue, setTimeoutValue] = useState<any>(null);
   const [loading, setLoading] = useState(false);
   const [fetched, setFetched] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
-  const [statsFilter, setStatsFilter] = useState<TFilterOrUndefined>({});
+  const [filter, setFilter] = useState<TFilterOrUndefined>({});
   const operator = "in_list";
 
   useEffect(() => {
+    // Only fetch values if there is values from a pre-defined filter
     if (!fetched && values.length !== 0) {
       fetchValues();
     }
   }, [values]);
 
   useEffect(() => {
-    const nextFilter = removeAttributeFromFilter(
-      generateFilter(zone, componentId, true),
-        attribute,
-    );
-    if (filterHasUpdated(setStatsFilter, statsFilter, nextFilter)) {
+    const nextFilter = generateFilter(zone, componentId, false, false);
+    if (filterHasUpdated(setFilter, filter, nextFilter)) {
       setFetched(false);
-      setData([]);
+      resetFiltersBelow({ id: componentId, zone: zone });
+      setZone({ ...zone });
     }
   }, [zone]);
+
+  useEffectUpdate(() => {
+    // Reset fetched status on filter change to ensure new values are fetched on next dropdown open
+    setFetched(false);
+  }, [filter]);
+
+  useEffectUpdate(() => {
+    // Reset data options when fetched is set to false to prevent showing incorrect options on dropdown open
+    if (!fetched) setData([]);
+  }, [fetched]);
 
   useEffect(() => {
     errorMessage && PopUpMessage({ message: errorMessage, type: "error" });
@@ -78,7 +87,7 @@ export function FilterMultiSelect(props: IFilterInput) {
         group_by: attribute,
         stats_fields: "",
         stats: "",
-        object_filters: JSON.stringify(statsFilter),
+        filter: JSON.stringify(filter),
       }).toString();
       dataSource
         .custom({
@@ -88,45 +97,42 @@ export function FilterMultiSelect(props: IFilterInput) {
         .then((res: any) => {
           const statsValues = res.data.meta.stats;
           setData(statsValues.map((item: any) => item.key[attribute]));
-          setLoading(false);
-          setFetched(true);
           updateDropdownText("No results found");
         })
         .catch((error: any) => {
           setErrorMessage(
             "Error fetching unique values for " +
-              attribute +
-              " in " +
-              objectType +
-              ". " +
-              error.message +
-              ".",
+            attribute +
+            " in " +
+            objectType +
+            ". " +
+            error.message +
+            ".",
           );
           console.error(error.message);
+          updateDropdownText("Error fetching values");
+        })
+        .finally(() => {
           setLoading(false);
           setFetched(true);
-          updateDropdownText("Error fetching values");
         });
     }
   };
 
-  filterListener(
-    {
-      attribute: attribute,
-      componentId: componentId,
-      operators: [operator],
-      zone: zone,
-      setValue: setValues,
-      setExists: setExists,
-      setNegate: setNegate,
-      setDisabled: setDisabled,
-      emptyValue: [],
-      zoneToValue: (filterValue: any) => {
-        return filterValue;
-      },
+  filterListener({
+    attribute: attribute,
+    componentId: componentId,
+    operators: [operator],
+    zone: zone,
+    setValue: setValues,
+    setExists: setExists,
+    setNegate: setNegate,
+    emptyValue: [],
+    zoneToValue: (filterValue: any) => {
+      return filterValue;
     },
-    [zone],
-  );
+    onlyUpdateMyValues: true,
+  });
 
   const onFilter = (input: string[]) => {
     setValues(input);
@@ -134,7 +140,7 @@ export function FilterMultiSelect(props: IFilterInput) {
     clearTimeout(timeoutValue!);
     setTimeoutValue(
       setTimeout(() => {
-        setFilter({
+        setFilterInput({
           operator: operator,
           value: input,
           negate: negate,
@@ -151,7 +157,7 @@ export function FilterMultiSelect(props: IFilterInput) {
   const onExists = (ex: boolean) => {
     setExists(!ex);
     setValues([]);
-    setFilter({
+    setFilterInput({
       operator: "exists",
       negate: negate,
       exists: !ex,
@@ -164,7 +170,7 @@ export function FilterMultiSelect(props: IFilterInput) {
 
   const onNegate = (ng: boolean) => {
     setNegate(!ng);
-    setFilter({
+    setFilterInput({
       operator: exists ? "exists" : operator,
       value: values,
       negate: !ng,
@@ -183,7 +189,7 @@ export function FilterMultiSelect(props: IFilterInput) {
   };
 
   const setDropdownText = () => {
-    if (!fetched) updateDropdownText("Loading values...");
+    if (!fetched) updateDropdownText("Loading options...");
     else if (errorMessage !== "") updateDropdownText("Error fetching values");
   };
 
@@ -193,7 +199,6 @@ export function FilterMultiSelect(props: IFilterInput) {
         block
         data={data}
         placeholder={rename}
-        disabled={disabled}
         value={values}
         setValue={onFilter}
         loading={loading}
@@ -208,7 +213,6 @@ export function FilterMultiSelect(props: IFilterInput) {
         onNegate={onNegate}
         exists={exists}
         onExists={onExists}
-        disabled={disabled}
       />
     </div>
   );