+++ CHANGELOG.md
@@ -1,11 +1,25 @@
 <!--
-SPDX-FileCopyrightText: 2023 Genome Research Ltd.
+SPDX-FileCopyrightText: 2026 Genome Research Ltd.
 
 SPDX-License-Identifier: MIT
 -->
 
 # Changelog
 
-## tol-ui 3.2.3 26-03-2026
+## tol-ui `3.4.0`
+02-04-2026
+- Multi-select filter options are now filterable (TOLP-8099)
+- `generateFilter` now reverts to the defaultFilter if no other filters are present (at a attribute and operator level)
+
+## tol-ui `3.3.0`
+26-03-2026
+
+- `More` & `Unknown` sections of the Sunburst visualisation are now clickable (TOLP-7103)
+- Deal with the 0 vs null/undefined DataPoint issue (TOLP-9664)
+- Removed max-width for content on SmartApp
+- Date picker for dates on editable cells (TOLP-9645)
+
+## tol-ui `3.2.3`
+26-03-2026
 
 - Added changelog file to track tol-ui version updates and changes